+++ EcoSimCMAKE/Order.h
@@ -5,15 +5,15 @@ class Pop;
 class Order {
 private:
 	int quantity;
-	int price;
+	double price;
 	std::string good;
 	Pop* owner;
 
 public:
 	int getQuantity();
-	int getPrice();
+	double getPrice();
 	void setQuantity(int newQuantity);
 	std::string getGood();
 	Pop* getOwner();
-	Order(int quantity, std::string good, int price, Pop* owner);
+	Order(int quantity, std::string good, double price, Pop* owner);
 };
